Consolidate Python code and update session utilities
