+++ README.md
@@ -1,1 +1,29 @@
-# EyeHeadCoupling
+# Eye-Head Coupling Analysis
+
+Tools for studying the relationship between eye and head movements in rats and tree shrews.
+
+## Directory Structure
+- `MATLAB/` – legacy MATLAB code.
+- `Python/` – all Python analysis code, utilities and notebooks.
+- `data/` – session configuration files and sample data.
+
+## Setup
+1. Clone the repository
+   ```
+   git clone https://github.com/SarvestaniLab/EyeHeadCoupling.git
+   cd EyeHeadCoupling
+   ```
+2. Create the conda environment
+   ```
+   conda env create -f Python/EyeHeadCoupling.yml
+   ```
+3. Activate the environment
+   ```
+   conda activate EyeHeadCoupling
+   ```
+
+## Usage
+- Run analysis scripts from `Python/analysis/`.
+- Launch Jupyter notebooks from `Python/notebooks/`.
+
+Contributions are welcome via pull requests.
